backend/requirements.txt
fastapi
uvicorn
yt-dlp
pytube
openai
whisper
torch
python-dotenv
